WIP fixed more tests
