+++ apps/coscrad-frontend/src/components/nav-menu/nav-menu-item.tsx
@@ -3,16 +3,15 @@ import { Link } from 'react-router-dom';
 import { NavItemInfo } from './nav-menu';
 
 export interface NavItemItemProps {
-    index: number;
     navItemInfo: NavItemInfo;
     handleClose: () => void;
 }
 
-export const NavMenuItem = ({ index, navItemInfo, handleClose }: NavItemItemProps) => {
+export const NavMenuItem = ({ navItemInfo, handleClose }: NavItemItemProps) => {
     const { link, label } = navItemInfo;
 
     return (
-        <MenuItem onClick={handleClose} key={index}>
+        <MenuItem onClick={handleClose}>
             <Link to={link}>{label}</Link>
         </MenuItem>
     );
